Add maru-hammerhead_defconfig for Marshmallow
This uses stock AOSP defconfig as a base and adds the minimum set of
options for LXC containers.

A key difference between this defconfig and the old one for L is that we
drop CONFIG_USER_NS since it doesn't compile properly on this branch.
This isn't a problem since user namespaces are only needed for
unprivileged containers in LXC; Maru OS only uses privileged
containers at the moment.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>